Added initial Vagrant configuration for provisioning Ubuntu with Titan installed.
There are a number of things not finished but this is a great jumping off point. It would be nice if we could setup a shared fixtures directory for developing against titian using different data.

The titan module should be broken out of this project and pulled in via submodule so that we can publish it to puppet forge.

"apt-get update" will run everytime and should really only be run when packages are required to be installed.

Of note is that we're not using the latest release of titan (3.2) due to [a bug](https://groups.google.com/forum/#!msg/aureliusgraphs/NSJOQDl13Gc/mdee0nUDQRMJ).